PUTWorkbench 1.3
Feature addition:
***************************
Feature: Added option to use random attribute combinations generation
for certain use-cases.
CLI change: Addition of new (optional) switch -rand to "advise" usage of
the new feature
UI change: Addition of a checkbox at the "Run Tasks" page to opt for the
new feature.
***************************

Bug fix:
***************************
Bug: Choosing "SGD" classifier with multi-class dataset throws an
Exception.
Fix: Added a compatibility check method that is invoked to see if the
selected dataset and parameters are compatible or not, flag the user
about any incompatibility promptly and gracefully exit.
***************************