PUTWorkbench 1.31
Bug fix:
***************************
Bug: The random attribute generation method may run into an infinite
loop if the number of combinations to generate are more than possible,
after considering privacy exceptions.
Fix: Added a time-out mechanism to check such cases which checks
periodically for progress in the generated pool, and stops the
generation if it has been stuck for a considerable amount of time.
***************************