PUTWorkbench 1.5
Feature addition:
***************************
Feature: Added a Verifier tool for verifying (sub)sets of results of an
experiment over a different size of the original dataset.
CLI change: Addition of a new tool "Verifier", part of both the
CLI as well as UI builds.
***************************

Tweaks:
***************************
Changed the heading of the metric "Area under PR Curve" to "a_PR"
instead of "a_PRC".
***************************